[pm] pm_publish_archive
This change implements a new gn template: `pm_publish_archive`.
The template follows the same procedure as `pm_prepare_publish`
combined with `pm_publish`, but produces a zip file as output,
rather than a repository directory of files. This template can be
used in combination with the `zip_resource` gn template to include
a fuchsia packages TUF repository in a fuchsia package. This is
useful, for example, to implement hermetic package delivery tests
that interact with an on-device package server.

Bug: 88453

Change-Id: I11e24cf429d6ba570af3e4381de670530ffc2030
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640725
Fuchsia-Auto-Submit: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Allison Pearce <ampearce@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>