[fidl] Template NaturalMessageEncoder on Transport
In an upcoming CL, NaturalTransport will need to store handle metadata
that is specific to a given transport. In order to facilitate this, this
CL templates NaturalMessageEncoder on Transport.

NaturalBodyEncoder is not templated because it should be ok for the
forseeable future to use the ChannelTransport with it.

Change-Id: I62eddf28ad3b6d9133ba27be878cbfdce5868b1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647925
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>