[fasync] Add fasync::scope
fasync::scope is an important mechanism for lifetime management so that
futures don't use (for example) dangling references after the referents
have died. One places an fasync::scope in the same scope as such
references or other resources and the relevant futures (once wrapped)
will automatically be canceled when the fasync::scope is destroyed.

Test: fasync-tests
Change-Id: I9c4427cd727e7b7e9b55cd381dc4784672a1f11b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627373
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Mark Schott <schottm@google.com>