[zircon][ldso] Avoid infinite recursion in dlsym
It's possible to infinitely recurse in find_sym_for_dlsym if DSOs
have circular dependencies. This dynamic linker code will be
eventually replaced. A hack we can do now to allow clang to roll is
dynamically check which DSOs we've visited already and if we hit a
cycle then we immediately fail the lookup.

Bug: 93847
Change-Id: Ie12707c2419df76881babf81f5dbc5cb550a2af0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647643
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Petr Hosek <phosek@google.com>