[build] Make hermetic action failures less verbose
If a build step that's wrapped by the action tracer fails in a generic
way (not related to action tracing) the error message would previously
include a stack trace from the action tracer script in addition to the
forwarded stdio of the failed command, e.g.:

  <full action tracer command line>
  <stdout/stderr of failed command>
  Traceback (most recent call last):
    File "../../build/tracer/hermetic_inputs_action.py", line 59, in <module>
      sys.exit(main())
    File "../../build/tracer/hermetic_inputs_action.py", line 48, in main
      subprocess.check_call(cmd_args)
    File "../../prebuilt/third_party/python3/linux-x64/lib/python3.8/subprocess.py", line 364, in check_call
      raise CalledProcessError(retcode, cmd)
  subprocess.CalledProcessError: Command '[<cmd>]' returned non-zero exit status 1.

(where <cmd> is the command line that's passed to the action tracer, and
is redundant because it's already included in the full action tracer
command line).

This just adds meaningless noise to the error message and makes it
harder to determine the actual cause of the failure, so we will now no
longer display that stack trace.

Change-Id: I15a251d71531a765994b7be70d95b2185907b9ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649623
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>