[fshost] Try to give at least 10% to /data
When we format a custom filesystem, we arbitrarily give it 24MiB
(100MiB for f2fs), which is a fairly small allocation, especially if the
disk is substantially bigger.  For some filesystems (fxfs, f2fs), this
is the final partition size, as they cannot grow.

Tweak the algorithm to give the greater of 10% of the disk or 24MiB.

Bug: 93060
Change-Id: Ia0b0b7312729fef82627f872d3a9b637495c805a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649192
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>