[ffx] Add UpdatedTarget event to improve target waiting.
This adds an UpdatedTarget event to improve target waiting behavior,
specifically if awaiting a manually added target for a specific nodename
that is known a priori.

If we manually add a target and then run the wait_for_target function,
we will receive NewTarget events for each new target seen (including
synthesized events). However, if we then make an RCS connection and
identify the target, we will not succeed in waiting for this target as
there is no indication that it has been updated.

Bug: 94317
Change-Id: Ifae46f6e11f1299b6763c5719b53fba29a871828
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649941
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>