[ffx][config] Fix tests overwriting user settings
ffx config was designed such that if it was run in a test, it would only
write to temporary data. While this was true for the environment config,
ffx_config::check_config_files would check if the environment contained
a path to a user config. If not, it would set it to the real default
user config. This would result in tests writing to the real user config.

To avoid this, this patch changes initialization of the test config to
create a temporary user config file, and point the environment at it.

Change-Id: Ibc449c10ce256f536dd7252ac8ea60a85130ae91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650061
Reviewed-by: Dangyi Liu <dangyi@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>