[audio_core] Add SilencePaddingStream
EffectsStage and MixStage need to "ring out" when a stream goes silent.
For example, a windowed resampler needs at least one full window of
silence to avoid a sudden discontinuity before the silence.

To implement ring out in a generic way, this CL adds
SilencePaddingStream. Each SilencePaddingStream wraps a source stream S.
When a gap is noticed (i.e. when S.ReadLock() returns std::nullopt), the
SilencePaddingStream returns a buffer containing a configurable amount
of explicit silence. For example, given:

  PacketQueue -> SilencePaddingStream -> MixStage

If the last packet has frames [20,30) and the resampler's window size is
5, then after the PacketQueue returns that last packet, the
SilencePaddingStream will follow with a silent buffer in frames [30,35).

A future CL will wrap the source streams for EffectsStage and MixStage.

Bug: 50669
Bug: 73306
Change-Id: I5d1b69cba728a2e517aa7e9c59f621826b166094
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640117
Reviewed-by: Martin Puryear <mpuryear@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>