[scenic] Removes the command line flag that forces flatland use
The flag was introduced during the transition from the gfx
APIs to the Flatland APIs use for graphics, to ensure that
the tests always execute using Flatland.

Now that it seems the Flatland API is the default and we
are not going back, let's remove this unneeded bit of
complexity.

Fixed: 92930
Change-Id: I5483553105a309acbac8b45fa43fbd3a8fdd5763
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650503
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>