[audio][virtual_audio] virtual_audio cleanup
The virtual_audio driver receives a bit of code cleanup --
largely conversion of DEBUG_ASSERTs and return-error to ASSERTs.
The thinking is (1) for aspects that we CAN control, we do not
intentionally configure things to cause errors, and (2) for
aspects that we cannot control, we want to surface problems
as soon as possible.

Change-Id: Ie1cf3b446e586eea31581c4f073366654d9f455a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/579460
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>