[input] Adds a display ownership handler to the input pipeline
This commit allows us to stop processing input when the user jumps away
from the product, as in the case of jumping to and back from
the virtcon.

While this commit will multiplex virtcon and product, I'm still not 100%
pleased with the way it is structured. But to resolve that part
completely, I'll need to do some additional work that this commit does
not need to wait on.  Hence, added some TODOs in the change and sending
this one for review.

Bug: 70249
Change-Id: I78aea299758132e85658822270fb130a95e5ebe9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648201
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>