[omaha-client] mock-omaha-server supports multiple apps
Mock OmahaServerBuilder and OmahaServer should accept vectors of
OmahaResponse, Hash, Option<UpdateCheckAssertion> objects in order to
support multiple apps in integration tests. This CL also verifies that
the number of apps in the request matches the number of apps configured
in the mock object, and adds a unit test for the ctor. Mock OmahaServer
also now accepts expectations for (and checks) app_id and version.

Change-Id: Id094ba9c8795f1591bd9f2d86bd988b779d8f9d2
Bug: 92396
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/649942
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: James Buckland <jbuckland@google.com>