[mdns] establish pattern for nonce service implementations
This CL moves the implementation of the ServiceInstanceResolver
service out of MdnsServiceImpl and into its own class, each instance
of which serves a single client.

The FIDL rubric requires that, when a client makes an avoidable error
(such as passing an invalid argument), the service implementation
should log a message and close the channel. This prevents us from
using any service implementation that serves more than one client
(because we have no way of knowing which connection to close).

Here we establish a pattern for nonce service implementations. The
ServiceImplManager class manages lifetime for service implementation
instances. The manager employs a deleter callback, passed to the
instance's constructor, which the instance can use to request
deletion. ServiceImplBase is a base class for the service
implementations handling binding, deletion, and connection errors.
It also supplies the instance with access to the Mdns class, which
is an in-proc API that the instances call do get work done.

This CL also changes ServiceInstanceResolver to handle client errors
as described. There are currently no clients out-of-tree, so this can
be done without breaking.

Test: no behavior change, unit test added

Change-Id: I743fa3268a0ddf1162b3df7fe996d3bd95b040a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650807
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>