[check-licenses] Results package
This handles saving NOTICE files, printing a summary to the command
line, and producing a manifest / summary of all the license texts that
were found in the build. Basically, all outputs for the tool should come
through this package.

We can move this back up to the base check-licenses package in the
future if it makes more sense.

Also added a driver file to check-licenses to make it easier for me to
test the v2 codepath. This code is still not being run by default.

Change-Id: I6092f8aa742987313afb8bcd092594f61d3c409c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650826
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>