[lowpan-ot-driver] Skeleton support for SRP Advertising Proxy.
This is not the full, working version of the SRP Advertising proxy---
it is missing the parts that interface with the mDNS responder service.
However, this change does allow the SRP service to work properly, which
is needed for testing.

Change-Id: I71a3c0ba4713bb128e9e470e8ce103cc4281776b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648784
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Jiaming Wang <jiamingw@google.com>
Fuchsia-Auto-Submit: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>