[fxfs] Add new metrics module
Provides a cross-platform unintrusive way of recording metrics for fxfs.

Metrics will be reported via Inspect under the fs.detail node. There are
no plans to support metrics collection on host builds yet, but the code
can still be compiled on host builds.

The interface of the metrics are almost identical to those of the
various property types in `fuchsia_inspect`, and for the most part can
act as drop-in replacements. This will allow for use of this crate
to be replaced with `fuchsia_inspect` once it supports cross-platform
compilation (fxbug.dev/94066).

Add metrics for max transaction size and completed transactions.

Bug: 94075
Test: Manually validated output by running `ffx inspect show
bootstrap/fshost:root/fxfs`.

Change-Id: I8e75b370d8020d71d49ceda3cf18817148b06d24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647551
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>