[element_manager][annotations] Introduce WatchAnnotations
This CL introduces the `WatchAnnotations` FIDL protocol method, and
implements the function for `felement::AnnotationController` and
`felement::Controller`.

The `watch_annotations` function utilizes the `async_utils::hanging_get`
API to perform a hanging-get so that clients can watch for annotation
updates.

A `WatchSubscriber` is initialized for each client stream to keep track
of the client state in regards to annotation updates. When
`watch_annotations` is called for the first time by the client, or if
the annotations have changed since the client's last request, the
current set of annotations are returned immediately. Otherwise, the
request will "hang" and a response will not be returned until the
annotations have been updated.

The reason for this change is to provide a mechanism for a client to
synchronize its annotations set with updates. This provides feature
parity with `StoryPuppetMaster.WatchAnnotations` in Modular for
`GraphicalPresenter`.

Bug: 56594
Change-Id: Ib2afe98d80a604cbb531175a359f8fabfddf97e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647783
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>