[input] Reworks the display ownership handling
I wasn't all too happy with the previous approach to handling events
because it didn't handle display ownership loss and gain at the time the
change happens, but rather at the time the next input events comes in.o

I reworked the approach, centralizing all event processing in a single
loop, and using async channels to funnel events in and out of the
handler.

What remains is a rename of the struct and files to not use the word
"handler", since in the input pipeline it has a very specific meaning.
I opted to rename in a separate change to simplify the resulting diff.

Bug: 70249
Change-Id: I4689eb52d035064c79bd67a3f2e9038960c11946
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650103
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>