[blobfs] Remove kDefaultJournalBlocks, use kMinimumJournalBlocks.
Both constant were the same value because the minimum value was quite
large to accommodate the largest operation. This removes all the logic
that tries to intelligently shrink/grow the journal size based on space
available, since it was previously making a larger journal for smaller
partition sizes. This unexpected behaviour was creating a mismatch in
the expected size of the filesystem in the host tool mkfs.

Had to update one test's values since it seemed to be just barely
passing due to this miscalculation.

Fixed: 94138
Change-Id: I0c2d7ca96d226ee0b95406152447b975069479f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651008
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>