[workstation][chromium] Deflake workstation_chrome_advanced_smoke_test
The test started to fail at opening the app launcher.
https://ci.chromium.org/ui/p/turquoise/builders/global.ci/sucrose_eng.chromebook-x64-debug-e2e/b8820983292116563057/overview
This isn't observed in local test, but is on the builders. I added the logic to repeat opening the app launcher and look for the Chromium app entry for a fix time duration as the flaky might be coming from the delay in completing a task on builders.

Bug: 94042

Change-Id: I4b815b98dfe46143606c2c7edcecb5a0767a0845
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/651581
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Yeonhee Lee <yhlee@google.com>