[bt][hci] Remove unused Rust HCI library
The HCI Transport library was meant to implement HCI transport drivers
without duplicating code.  Unfortunately Rust is not a supported DDK
language and this library has no maintainer for now, as well as
outstanding Leak Sanitizer issues.  Remove it for now and re-evaluate
whether to use Rust for this functionality in transport drivers.

Bug: 91394
Fixed: 91768

Test: code removal, fx build and CQ
Change-Id: I96fb9f50844ff83d2d0fba84a9f26865f58d32da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651614
Reviewed-by: Joshua Conner <joshconner@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>