[vm][test] Handle VMO resize
Remove ZX_VMO_RESIZABLE from the default creation option in pager core
tests.

Modify the VMO resize helper to also change the mapping when the VMO is
resized. This is needed to be able to access the VMO safely after a
resize.

No change in behavior with this CL. These changes are needed for tests
that will be added in following CLs.

Test: core-pager
Bug: 94371
Change-Id: I6ce4d4d3676d00a742bfc6c347b23272fdf14b25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650841
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>