[kernel][vm] Introduce supply_zero_offset
Pager-backed VMOs get a new supply_zero_offset, which is the offset
beyond which new page requests are implicitly fulfilled with zero pages
instead of forwarding a read request to the user pager.

The supply_zero_offset is used as follows:
 - The supply_zero_offset is initialized to the VMO size on creation.
 - A resize up leaves the supply_zero_offset unchanged.
 - A resize down sets the supply_zero_offset to min(supply_zero_offset,
   size).
 - The kernel supplies zero pages for the range [supply_zero_offset,
   size).

This CL introduces the supply_zero_offset and updates it on resizes.
Supplying zero pages beyond supply_zero_offset will be implemented in a
following CL.

Test: kernel-unittests
Bug: 94371
Change-Id: I800e89162bbcd6bfd6cd84531981d957168cd037
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650102
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>