[input] Renames DisplayOwnershipHandler to DisplayOwnership
Also fixes all other places affected by this rename.

The struct is renamed because the word 'handler' refers to the trait
`InputEventHandler` in the input pipeline. But this struct does not
implement it because it must use the channel interface to process
events that aren't just input events.

Fixed: 70249
Change-Id: Ica35e3ed80fcad96655691c311639abebcf64dc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650201
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>