[lowpan][openthread_fuchsia] Fix for Platform UDP Sockets.
There were two problems with the existing Platform UDP sockets:
The first is that the socket was being opened as an IPv4 socket
instead of an IPv6 socket. The second was that in error conditions
when sending we were double-freeing the message. This CL addresses
both problems.

Change-Id: I7734464025ce08ed193c6b8e111fbc6b8a65080b

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651722
Reviewed-by: Jiaming Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Fuchsia-Auto-Submit: Robert Quattlebaum <rquattle@google.com>