[Recovery] Fix button race conditions.
Button presses were not recorded unless matched in a match
statement. They are now recorded and propagated as a plain
button press.

There was a race condition when in waiting mode when a cancelled
event could get through because of a background task in main.
This is now ignored.

Rustfmt did a lot of reformatting.

Bug: b/215578671
Fixes: b/215578671
Test: New test added, fails before fix
Test: One test removed, it is no longer relevant
Test: All tests run (the same three fail that failed before)
Test: Verified that it no longer panics when both buttons released.
Change-Id: I41d7e0620814d8d0c6162e4736b238618dc0c2bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634723
Fuchsia-Auto-Submit: Kevin Twidle <kpt@google.com>
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>