[size-checker] Write the report to file.
By default the size_checker tool writes the report to stdout. When `--report-out path` is specified the report is written to a file, and only errors are written to standard stream.

This makes possible to run the tool as part of the build system, where nothing should be printed when the tool runs without error.

Change-Id: I36dd80ff0859b6ad4e1779d2010fcef2a93a3185
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642801
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Pierre Labatut <plabatut@google.com>