[build][rust] refactor common rustc_* code into rustc_artifact
For now, only features and edition logic are implemented in the helper,
but enforce_source_listing, clippy, and with_unit_tests might be good
candidates for extraction in the future.

This change should be a no-op refactor, not changing anything about the
generated targets.

Change-Id: I8e9d4cfc1df1be52cab33114c8a844e3824a8b25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650882
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>
Commit-Queue: Dan Johnson <computerdruid@google.com>