[bsdsocket_test] Stringify cmsg type
Instead of having to manually translate cmsg level/type to a string, use
a macro to do it for us. It has the benefit that we do not need to add a
new switch case every time a new cmsg is added.

Change-Id: Ic6f636c16ae6d2a0d9d955395dbd04c28db976dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/651421
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Arthur Sfez <asfez@google.com>
Fuchsia-Auto-Submit: Arthur Sfez <asfez@google.com>