[system-update-checker] Fix flaky tests
By virtue of the test fakes running quickly, the tests that verify that
the status was committed succeeded. By adding a timeout, these tests
failed because they check that an update was started when they should
check that an update has completed. Edited to first check that the
update is in a terminal state and then verify that the CommitQuerier
makes its fidl calls to be more resilient to the speed at which the
test runs.

Fixed: 91985

Change-Id: Ibcdb7f9e54c75f25e478983cb8009dcea5fa33d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650353
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>