[intl] Replace the use of sr-RS with en-US in an example
The sr-RS locale is still evolving, making change detector tests like
the ones edited in this change brittle when ICU versions are being
updated.

Change detector tests aren't the best approach for L10N-ized tests,
but in this case the impact is small and infrequent such that it
doesn't really seem worthwhile to change.

We really just want to verify that the language is changed when
the locale changes and that is it.  For that reason use locales
with more mature data in the test, such as en-US.

This should also allow us to update the ICU library use in
Fuchsia from 69.1 to 70.1.

Bug: 92408
Change-Id: I54c0a539bed28cad3ec48a310797c2258d06bb58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652254
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>