[rust-3p] Actually preserve line endings
According to the [git docs], `text=false` is not a valid setting. This
results in the setting being treated as unspecified, which means git
will use the setting for `core.autocrlf`.

This fixes this, by instead using `-text`, which informs git to not
attempt any end-of-line changes.

In addition, this corrects the vendored gpt crate, which is using crlf
line endings.

[git docs]: https://git-scm.com/docs/gitattributes#_checking_out_and_checking_in

Change-Id: I90989d0e181376a3f43693d4ee3b21a46de91fda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652622
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>