[fidl] Natural encoder validation interface
This introduces a mechanism to perform valdation in the natural type
encoding process. Validation is only implemented for bools for now, but
is intended to extend to all types. This CL is mostly being sent out to
confirm the direction to go in and to create a basis for adding further
validation checks.

Change-Id: I7cc8e39de12f86dd8b2d009eda55a986792e78df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650941
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>