[fuchsia-zircon-tests] Port fuchsia-zircon-tests to V2.
This change ports fuchsia-zircon-tests to a V2 test. We are attempting to remove
all instances of fuchsia.boot.RootResource from appmgr so we no longer
need to route the sensitive capability down that far into the system.
Note even though only tests use this capability due to the legacy of appmgr
we still have to route the capability down in production so we would
like to clean that up correctly.

Change-Id: I200a4b9ac73a5ad0238078f5637b2f76d55fe4de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648388
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Ben Wright <benwright@google.com>