[amlogic][nand] Poll chip status instead of waiting before page read
Add an option to poll the chip status register to make sure it is safe
to read instead of just waiting 25us. This duration is close to the
maximum page read time for Toshiba NAND chips, which may explain the
intermittent failures that are suspected to be happening in the field.
Micron is switched to polling as well.

Test: Ran iotime read on FVM partition on Astro
MULTIPLY: aml-rawnand-unittest: 500
Change-Id: Ie048a0e40a550019ec3c3d33631991a2d197b804
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650003
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>
(cherry picked from commit 99e025142922860161f5c08b4b8be73215034b52)