[scenic][flake] Flake fix for observer integration test.
The observer integration test assumes that the number of view tree
snapshots generated are deterministic which is incorrect and depends
upon the frame scheduler. This CL modifies the way assertions are done
in the test to do away with this assumption.

Fixed: 92859

Multiply: observer_integration_tests: 2000

Change-Id: Iaa9ccc65389d35f155711c832411d71b3ddc7ff5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652463
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: Shivesh Ganju <shiveshganju@google.com>