[vm-stress-test] Port vm-stress-test to V2.
This change ports vm-stress-test to a V2 test. We are attempting to remove
all instances of fuchsia.boot.RootResource from appmgr so we no longer
need to route the sensitive capability down that far into the system. Note
even though only tests use this capability due to the legacy of appmgr we
still have to route the capability down in production so we would like to
clean that up.

Testing:
I tested the change on a qemu-core.x64 and it started up and ran correctly.

Change-Id: I492b32c46218df10d92e1a8109dd82cd74a8d073
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648382
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Ben Wright <benwright@google.com>