[kernel][vm] Fix overflow in ExpandContentIfNeeded
When a stream operation wants to expand the content size, we decide
whether to resize the VMO by computing the new required size as
ROUNDUP(new content size, PAGE_SIZE). If the roundup overflows, we will
incorrectly infer that the VMO does not need resizing, and will end up
updating the content size without having resized the VMO. This is
incorrect as the content size should never be larger than the VMO size.

Test: core-stream
Bug: 94454
Change-Id: I010e426114ba28ef39d55a87f14fd04560f78ad2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652383
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>