[virtualization] Add batch operations to BlockDispatcher
This change extends the BlockDispatcher interface to provide operations
to read/write batches of requests and the virtio_block device is updated
to submit all the request in a single chain as a single batch.

None of the dispatchers are updated to provide custom batch operations
yet. Instead we just add a default implementation of ReadBatch and
WriteBatch that call the existing ReadAt and WriteAt methods in a loop.

Bug: 93355
Change-Id: Id6359a1bf9d1efbe22b5c45443aef8fe9d3ad818
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652745
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>