[fidl][cpp] Design doc for unified bindings
Since we would be referencing sections of the design doc in CLs quite a
bit, this CL makes the design doc public, and applies some wordsmithing
and editorial adjustments to make it easier to read.

In doing so, some obsolete content were removed. Some more rationale was
given behind particular decisions.

As of now the doc is not written up to "the RFC standard"; its role is
closer to a living reference that we could point to in public, and it is
expected that parts would be amended as we learn more through the
iteration and user feedback of the bindings.

Test: checked for style violations via
    fx mdlint --enable all --root-dir src/lib/fidl/cpp/docs
Test: staged the doc on devsite

Change-Id: I6bf23afc97c3b80019f596a3568765bcd62e5237
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614030
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>