[media] Return status from call failure
A recent CL
Ic56ed1521003f3f382549e157158be0dfcf1fdd3
added an assertion on call failure, which may not be safe because it
crashes the process on failure. This CL updates the behavior to return
the status.

Let me know whether or not this is the correct way to handle a call
failure.

Change-Id: I6eaceb0d8d8cc35b04b9e13363e8c31ca335c8aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653922
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>