[fidl_driver] Tests for client and server reset
This CL adds tests that the correct callback is invoked by the server
when the client end is reset and likewise by the client when the server
end is reset.

This only covers the WireSharedClient case, but the WireClient case
can be added once WireClient calls are working properly. It is
arguable that it may not be needed, though, since the logic largely
shares the same code paths.

Multiply: fidl_driver_transport_tests (fuchsia): 100

Change-Id: I41b33505ca35fc7707b37ecaba8642688a375b7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650401
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>