[component][fuzzing] Add AsyncDeque
This CL adds a class that can be used to asynchronously send data from
one future to another.

As the first CL with promises and futures, it also adds some utilities
such as type aliases and a base class for unit test that include methods
for running promises and checking their results.

Bug: 92490
Change-Id: I083d091c7e1539fe7585440389412df5186cd3ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640101
Reviewed-by: Nick Cano <nickcano@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>