[flatland] Check display signal fences before SetLayerImage
This CL adds a workaround for the current restrictions of the display
API where the clients aren't allowed to reuse the same image in
consecutive configs. DisplayCompositor sends signal fences along with
each image and waits for them to be signaled before attempting to use
again. DisplayCompositor goes for vulkan fall back if the image cant
be used.

Fixed: 91770
Change-Id: Id69f10385e78c420cc1fb01b239d5bae1e05670a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652682
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Emircan Uysaler <emircan@google.com>