[build][dart] Default to AOT in release mode, and corrected GN template errors
The root cause behind the following bug is resolved, so the build
scripts should no longer force JIT for release builds.

Fixed: 64126

Also...

`dart_test_component()` and `flutter_test_component()` accept a
`build_cfg` variable from the invoker, but the GN syntax was wrong.
Builds failed when I tried to use it to force an individual test to run
in AOT mode (by supplying the dart_aot_runner in the test's
configuration).

Note, this is what I added to my `dart_test_component()` declaration,
which works, with the changes in this CL:

```gn
  # Fuchsia non-system tests do not have permissions for
  # `ambient_mark_vmo_exec` (per the current `component_manager_policy`),
  # which is required by the JIT runners, so build the test in AOT
  # mode, only, and serve the AOT runner from the test component's
  # environment.
  if (is_debug && !dart_force_aot) {
    # Note that `dart_aot_debug_build_cfg` does not work in the default,
    # non-release Fuchsia build because that config enables asserts, and
    # the current AOT VM apparently requires `no-asserts`.
    build_cfg = dart_profile_build_cfg
  }
```

Change-Id: I6b36ac0498764880f31c8cba57d59c1adc149cf4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653646
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>