[flatland] Rectangle Compositor turn color conversion on/off
Create two render paths in the RectangleCompositor for rendering
with and without color conversion. Having a renderpath that does
not use color conversion will be useful for when there is no color
conversion data turned on, and for when for future group opacity
related usecases where intermediary passes do not require and conversion.

Change-Id: I8e9e331cfca2cadff44854bd323c986a49de25e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/650505
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Chris Fontas <cfontas@google.com>