[vfs] Use non-discardable fidl::Result
fidl::Result will soon be marked [[nodiscard]], making it necessary to
use the value. For the sake of this CL, the result is cast to void,
keeping the existing behavior.

Note: the first version of this CL checked that the result's status
was ZX_OK, but this failed on CQ. It may be useful to follow up on
this CL and diagnose the reason for it.

Bug: 94739
Change-Id: I337e85b461922b7bc01b7f2bc357c4128c39d74e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653429
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>