[security][pkg] bad_signature_update test case
This change implements a security package delivery test that
verifies behaviour of applying an update that is legitimate in
every way except for the software vendor signature.

The test case is very similar to existing software update-based
tests, but expects the update to fail.

Bug: 88453

Change-Id: I9d23b24a779fdde981df2127e833fc3d7ef5f769
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/652111
Reviewed-by: Allison Pearce <ampearce@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>