[build] Replace rustflags target-feature=+crypto with +aes,+sha2
Something changed recently in rustc that causes it to generate warnings
when building for arm64 targets.

This appears related to https://github.com/rust-lang/rust/pull/87729
(though that PR is dated), which indicates that rustc is removing the
crypto target and replacing it with "aes or sha2".

This PR adds both, and removes crypto, for Rust only. (This does not
work with clang (cflags).

Change-Id: I438e2d3ae230fb512c7c864d775537ce4dd797f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654243
Reviewed-by: Chris Tam <godtamit@google.com>
Fuchsia-Auto-Submit: Rich Kadel <richkadel@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>