[bt][avrcp] Define production and testonly core shards
This change lays the groundwork for making the AVRCP component product
variable.

Define a core shard for AVRCP-only to be included in configurations that
support AVRCP.

Remove AVRCP's decl in core.cml so that it can be a product-variable
child of core.

Remove AVRCP's capability routes from appmgr as we can no longer
define static routes like these since AVRCP is product-dependent.

We expect A2DP and AVRCP to exist in all current configurations.
Therefore, it's sensible to include the AVRCP shard in the A2DP shard.
In the future, should the need arise for A2DP-only, we can define an
A2DP-only core shard.

Define a testonly core shard to expose the testing `PeerManagerExt`
capability route.

Bug: 86390
Test: fx build
Test: Manual, fx reboot and verify bt-avrcp, bt-a2dp startup OK.

Change-Id: I94a99db677da67dbd32bc84be00898feada0ec3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/653525
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>